Export RTE mockRangeForBoundingRect
