+++ .changeset/yellow-eels-tie.md
@@ -1,0 +1,5 @@
+---
+'@kaizen/components': patch
+---
+
+Export RTE mockRangeForBoundingRect helper function
