+++ packages/components/src/RichTextEditor/utils/commands/addMark.spec.ts
@@ -2,11 +2,7 @@ import { findByText, waitFor } from '@testing-library/dom'
 import { vi } from 'vitest'
 import { createRichTextEditor } from '../core'
 import { addMark } from './addMark'
-import {
-  mockRangeForBoundingRect,
-  simulateSelectionByText,
-  simulateTextInsert,
-} from './fixtures/helpers'
+import { mockRangeForBoundingRect, simulateSelectionByText, simulateTextInsert } from './fixtures/'
 import { testEditorState, testSchema } from './fixtures/test-state'
 describe('addMark()', () => {
   const onChange = vi.fn()
