+++ packages/components/src/RichTextEditor/utils/commands/fixtures/helpers.ts
@@ -5,38 +5,7 @@ import {
   type Transaction,
 } from 'prosemirror-state'
 import { findChildrenByType } from 'prosemirror-utils'
-//  eslint-disable-next-line import/no-extraneous-dependencies
-import { vi } from 'vitest'
 import { type ProseMirrorModel } from '../../prosemirror'
-/*
- ** This is used handle the JSDom type error issue you may encounter in testing
- ** See https://github.com/jsdom/jsdom/issues/3002
- */
-export const mockRangeForBoundingRect = (): void => {
-  vi.spyOn(document, 'createRange').mockImplementation(() => {
-    const range = new Range()
-
-    range.getBoundingClientRect = () => ({
-      x: 0,
-      y: 0,
-      bottom: 0,
-      height: 0,
-      left: 0,
-      right: 0,
-      top: 0,
-      width: 0,
-      toJSON: () => undefined,
-    })
-
-    range.getClientRects = () => ({
-      item: () => null,
-      length: 0,
-      [Symbol.iterator]: vi.fn(),
-    })
-
-    return range
-  })
-}
 
 /*
  ** Prosemirror cannot intuit a real `Selection` of content injected into the base test.
