+++ packages/components/src/RichTextEditor/utils/commands/fixtures/index.ts
@@ -1,0 +1,2 @@
+export * from './mockRangeForBoundingRect'
+export * from './helpers'
